Took should return int.MaxValue when value is greater than int
Add clear obsolete messages on Took int properties
Add JsonIgnore attributes to Took fields so that Json serializer ignores them when enumerating properties to serializes
Fix build warnings for usage of Took and check TookAsLong value is populated

+++ src/Benchmarking/Results.cs
@@ -15,7 +15,7 @@ namespace Benchmarking
 
 		public int IndexedDocuments { get; set; }
 
-		public IEnumerable<int> EsTimings { get; set; }
+		public IEnumerable<long> EsTimings { get; set; }
 
 		public Metrics Before { get; set; }
